% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_clear_subraster_cache}
\alias{rs_clear_subraster_cache}
\title{Empty the repaint-boundary sub-raster cache (FW4c) and reset its counters.}
\usage{
rs_clear_subraster_cache()
}
\description{
Empty the repaint-boundary sub-raster cache (FW4c) and reset its counters.
}
