% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\name{rs_lines}
\alias{rs_lines}
\title{Draw a polyline}
\usage{
rs_lines(scene, x, y, units = "npc", col = "black", lwd = 1, alpha = 1)
}
\arguments{
\item{scene}{A \code{\link[=rs_scene]{rs_scene()}}.}

\item{x, y}{Parallel coordinate vectors.}

\item{units}{Coordinate system; see \code{\link[=rs_rect]{rs_rect()}}.}

\item{col}{Line colour (\code{NA} for none).}

\item{lwd}{Line width (1 == 1/96 inch).}

\item{alpha}{Opacity multiplier in \verb{[0, 1]}.}
}
\value{
\code{scene}, invisibly.
}
\description{
Draw a polyline
}
