% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\name{rs_polygon}
\alias{rs_polygon}
\title{Draw a polygon}
\usage{
rs_polygon(
  scene,
  x,
  y,
  units = "npc",
  fill = NA,
  col = "black",
  lwd = 1,
  alpha = 1
)
}
\arguments{
\item{scene}{A \code{\link[=rs_scene]{rs_scene()}}.}

\item{x, y}{Parallel coordinate vectors.}

\item{units}{Coordinate system; see \code{\link[=rs_rect]{rs_rect()}}.}

\item{fill}{Fill colour (\code{NA} none, \code{NULL} inherit).}

\item{col}{Line colour (\code{NA} none, \code{NULL} inherit).}

\item{lwd}{Line width (1 == 1/96 inch); \code{NULL} inherits.}

\item{alpha}{Opacity multiplier in \verb{[0, 1]}; \code{NULL} inherits.}
}
\value{
\code{scene}, invisibly.
}
\description{
Draw a polygon
}
