% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\name{rs_rect}
\alias{rs_rect}
\title{Draw a rectangle}
\usage{
rs_rect(
  scene,
  x = 0.5,
  y = 0.5,
  width = 1,
  height = 1,
  units = "npc",
  fill = NA,
  col = "black",
  lwd = 1,
  alpha = 1
)
}
\arguments{
\item{scene}{A \code{\link[=rs_scene]{rs_scene()}}.}

\item{x, y}{Rectangle centre.}

\item{width, height}{Rectangle size.}

\item{units}{Coordinate system for the above: one of \code{"npc"}, \code{"native"},
\code{"mm"}, \code{"in"}, \code{"pt"}.}

\item{fill, col}{Fill and border colours (\code{NA} for none).}

\item{lwd}{Border line width (1 == 1/96 inch).}

\item{alpha}{Opacity multiplier in \verb{[0, 1]}.}
}
\value{
\code{scene}, invisibly.
}
\description{
Draw a rectangle
}
