% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\name{rs_render}
\alias{rs_render}
\title{Render a scene to a PNG file}
\usage{
rs_render(scene, path)
}
\arguments{
\item{scene}{A \code{\link[=rs_scene]{rs_scene()}}.}

\item{path}{Output file path.}
}
\value{
\code{path}, invisibly.
}
\description{
Render a scene to a PNG file
}
