% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_set_cvd_mode}
\alias{rs_set_cvd_mode}
\title{Set colour-vision-deficiency simulation for subsequent renders.}
\usage{
rs_set_cvd_mode(kind)
}
\arguments{
\item{kind}{One of "protanopia", "deuteranopia", "tritanopia",
"achromatopsia", or "" to disable.}
}
\description{
Set colour-vision-deficiency simulation for subsequent renders.
}
\keyword{internal}
