% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_set_glyph_bitmap_mode}
\alias{rs_set_glyph_bitmap_mode}
\title{Set the glyph-bitmap cache mode.}
\usage{
rs_set_glyph_bitmap_mode(mode)
}
\arguments{
\item{mode}{Integer: 0 = off, 1 = auto (threshold), 2 = on.}
}
\description{
Set the glyph-bitmap cache mode.
}
\keyword{internal}
