% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_set_profiling}
\alias{rs_set_profiling}
\title{Arm or disarm per-node render timing on this thread.}
\usage{
rs_set_profiling(on)
}
\arguments{
\item{on}{Logical.}
}
\description{
Arm or disarm per-node render timing on this thread.
}
\keyword{internal}
