% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_set_simplify_tol}
\alias{rs_set_simplify_tol}
\title{Set the path-simplification tolerance in device pixels (0 disables).}
\usage{
rs_set_simplify_tol(tol)
}
\arguments{
\item{tol}{Tolerance in device px.}
}
\description{
Set the path-simplification tolerance in device pixels (0 disables).
}
\keyword{internal}
