% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_svg_path}
\alias{rs_svg_path}
\title{Parse SVG path data into flattened rings.}
\usage{
rs_svg_path(d)
}
\arguments{
\item{d}{The \code{d} attribute of an SVG \verb{<path>}.}
}
\value{
List of \code{x}, \code{y}, \code{nper}, \code{closed}.
}
\description{
Parse SVG path data into flattened rings.
}
\keyword{internal}
