% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_take_node_times}
\alias{rs_take_node_times}
\title{Take the per-node render times (seconds) collected since the last take.}
\usage{
rs_take_node_times()
}
\description{
Take the per-node render times (seconds) collected since the last take.
}
\keyword{internal}
