% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vellum-package.R
\docType{package}
\name{vellum-package}
\alias{vellum}
\alias{vellum-package}
\title{vellum: A Low-Level Graphics Framework with a Rust Backend}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A grid-like low-level graphics system for R whose scene graph, unit/layout engine, and rendering run in a Rust backend.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-vellum/vellum}
  \item \url{https://r-vellum.github.io/vellum/}
  \item \url{https://schochastics.r-universe.dev/vellum}
  \item Report bugs at \url{https://github.com/r-vellum/vellum/issues}
}

}
\author{
\strong{Maintainer}: David Schoch \email{david.schoch@cynkra.com}

Authors:
\itemize{
  \item David Schoch \email{david.schoch@cynkra.com}
}

}
\keyword{internal}
